+++ tests/testthat/test-get_headlines.R
@@ -1,12 +1,12 @@
-test_that("get_headlines works", {
+test_that("get_headlines returns a data frame with one column", {
   # adding a small time delay to avoid simultaneous posts to the api
   Sys.sleep(3)
 
   # Get the news
-  output_headlines <- get_headlines("news.ycombinator.com")
+  output_headlines <- get_headlines("ycombinator.com")
 
   # What's returned
-  expect_is(output_headlines, "data.frame")
+  expect_s3_class(output_headlines, "data.frame")
   expect_equal(ncol(output_headlines), 1)
 
 })