+++ NEWS.md
@@ -1,3 +1,17 @@
+# newcatcheR 0.1.0
+
+* this update catches up to the changes made in the python package
+
+* updated database of websites and rss feeds
+
+* fixed code to match new database columns
+
+* new functions to get description and topics of URLs
+
+* removed `tld_sources()`, `filter_urls()` is a replacement
+
+* updated documentation
+
 # newcatcheR 0.0.2
 
 * make sure the package imports tidyRSS 2.0.2